+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 - MastoFE
+- SSH access mode
+- Gopher server
 
 ### Changed
 - Allow users to remove their emails if instance does not need email to register